rename Sidebar components in editor to omit confusion
